+++ tracdap-services/tracdap-svc-orch/src/main/java/org/finos/tracdap/svc/orch/service/JobExecutor.java
@@ -15,12 +15,13 @@
  * limitations under the License.
  */
 
-package org.finos.tracdap.common.exec;
+package org.finos.tracdap.svc.orch.service;
 
 import org.finos.tracdap.api.internal.*;
 import org.finos.tracdap.common.config.ConfigFormat;
 import org.finos.tracdap.common.config.ConfigParser;
 import org.finos.tracdap.common.exception.*;
+import org.finos.tracdap.common.exec.*;
 import org.finos.tracdap.common.grpc.ClientCompressionInterceptor;
 import org.finos.tracdap.common.grpc.GrpcChannelFactory;
 import org.finos.tracdap.common.grpc.ClientLoggingInterceptor;
@@ -45,7 +46,7 @@ import java.util.concurrent.Flow;
 import java.util.regex.Pattern;
 
 
-public class BatchJobExecutor<TBatchState extends Serializable> implements IJobExecutor<BatchJobState<TBatchState>> {
+public class JobExecutor<TBatchState extends Serializable> implements IJobExecutor<JobExecutorState<TBatchState>> {
 
     public static final Pattern RUNTIME_ENV_KEY = Pattern.compile("runtime\\.env\\.(\\w+)");
     public static final Pattern RUNTIME_PROP_KEY = Pattern.compile("runtime\\.((\\w+\\.)*\\w+)");
@@ -61,7 +62,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
 
     private GrpcChannelFactory channelFactory;
 
-    public BatchJobExecutor(IBatchExecutor<TBatchState> batchExecutor, PluginConfig executorConfig) {
+    public JobExecutor(IBatchExecutor<TBatchState> batchExecutor, PluginConfig executorConfig) {
         this.batchExecutor = batchExecutor;
         this.executorConfig = executorConfig;
         this.configParser = new ConfigParser();
@@ -80,17 +81,17 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
 
     @Override
     @SuppressWarnings("unchecked")
-    public Class<BatchJobState<TBatchState>> stateClass() {
-        return (Class<BatchJobState<TBatchState>>) (Object) BatchJobState.class;
+    public Class<JobExecutorState<TBatchState>> stateClass() {
+        return (Class<JobExecutorState<TBatchState>>) (Object) JobExecutorState.class;
     }
 
     @Override
-    public BatchJobState<TBatchState> submitJob() {
+    public JobExecutorState<TBatchState> submitJob() {
         throw new ETracInternal("Not implemented yet");
     }
 
     @Override
-    public BatchJobState<TBatchState> submitOneshotJob(TagHeader jobId, JobConfig jobConfig, RuntimeConfig sysConfig) {
+    public JobExecutorState<TBatchState> submitOneshotJob(TagHeader jobId, JobConfig jobConfig, RuntimeConfig sysConfig) {
 
         // If an error occurs before the batch state is created, then there is nothing to clean up
 
@@ -162,7 +163,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
 
             batchState = batchExecutor.submitBatch(batchKey, batchState, batchConfig);
 
-            var jobState = new BatchJobState<TBatchState>();
+            var jobState = new JobExecutorState<TBatchState>();
             jobState.batchKey = batchKey;
             jobState.batchState = batchState;
             jobState.runtimeApiEnabled = runtimeApiEnabled;
@@ -193,18 +194,18 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
     }
 
     @Override
-    public BatchJobState<TBatchState> submitExternalJob() {
+    public JobExecutorState<TBatchState> submitExternalJob() {
         throw new ETracInternal("Not implemented yet");
     }
 
     @Override
-    public BatchJobState<TBatchState> cancelJob(BatchJobState<TBatchState> jobState) {
+    public JobExecutorState<TBatchState> cancelJob(JobExecutorState<TBatchState> jobState) {
         jobState.batchState = batchExecutor.cancelBatch(jobState.batchKey, jobState.batchState);
         return jobState;
     }
 
     @Override
-    public void deleteJob(BatchJobState<TBatchState> jobState) {
+    public void deleteJob(JobExecutorState<TBatchState> jobState) {
         batchExecutor.deleteBatch(jobState.batchKey, jobState.batchState);
     }
 
@@ -214,7 +215,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
     }
 
     @Override
-    public RuntimeJobStatus getJobStatus(BatchJobState<TBatchState> jobState) {
+    public RuntimeJobStatus getJobStatus(JobExecutorState<TBatchState> jobState) {
 
         var batchStatus = batchExecutor.getBatchStatus(jobState.batchKey, jobState.batchState);
 
@@ -279,7 +280,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
         }
     }
 
-    private RuntimeJobStatus getStatusFromApi(BatchJobState<TBatchState> jobState) {
+    private RuntimeJobStatus getStatusFromApi(JobExecutorState<TBatchState> jobState) {
 
         var runtimeApiAddress = getRuntimeApiAddress(jobState);
         var runtimeChannel = (ManagedChannel) null;
@@ -306,7 +307,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
         }
     }
 
-    private void updateStatusFromLogs(BatchJobState<TBatchState> jobState, RuntimeJobStatus.Builder batchJobStatus) {
+    private void updateStatusFromLogs(JobExecutorState<TBatchState> jobState, RuntimeJobStatus.Builder batchJobStatus) {
 
         // If a log file is available, fetch it and look for a TRAC exception message
         // If the error log is not available, fall back on the basic batch status
@@ -365,12 +366,12 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
     }
 
     @Override
-    public Flow.Publisher<RuntimeJobStatus> followJobStatus(BatchJobState<TBatchState> jobState) {
+    public Flow.Publisher<RuntimeJobStatus> followJobStatus(JobExecutorState<TBatchState> jobState) {
         throw new ETracInternal("Not implemented yet");
     }
 
     @Override
-    public RuntimeJobResult getJobResult(BatchJobState<TBatchState> jobState) {
+    public RuntimeJobResult getJobResult(JobExecutorState<TBatchState> jobState) {
 
         var batchStatus = batchExecutor.getBatchStatus(jobState.batchKey, jobState.batchState);
         var batchCode = batchStatus.getStatusCode();
@@ -389,7 +390,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
         throw new ETracInternal(message);
     }
 
-    private RuntimeJobResult getResultFromApi(BatchJobState<TBatchState> jobState) {
+    private RuntimeJobResult getResultFromApi(JobExecutorState<TBatchState> jobState) {
 
         var runtimeApiAddress = getRuntimeApiAddress(jobState);
         var runtimeChannel = (ManagedChannel) null;
@@ -416,7 +417,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
         }
     }
 
-    private RuntimeJobResult getResultFromResultFile(BatchJobState<TBatchState> jobState) {
+    private RuntimeJobResult getResultFromResultFile(JobExecutorState<TBatchState> jobState) {
 
         var resultFile = String.format("job_result_%s.json", jobState.batchKey);
 
@@ -447,7 +448,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
         }
     }
 
-    private InetSocketAddress getRuntimeApiAddress(BatchJobState<TBatchState> jobState) {
+    private InetSocketAddress getRuntimeApiAddress(JobExecutorState<TBatchState> jobState) {
 
         var runtimeApiAddress = batchExecutor.getBatchAddress(jobState.batchKey, jobState.batchState);
 
@@ -470,7 +471,7 @@ public class BatchJobExecutor<TBatchState extends Serializable> implements IJobE
                 .newBlockingStub(clientChannel)
                 .withCompression(ClientCompressionInterceptor.COMPRESSION_TYPE)
                 .withInterceptors(new ClientCompressionInterceptor())
-                .withInterceptors(new ClientLoggingInterceptor(BatchJobExecutor.class));
+                .withInterceptors(new ClientLoggingInterceptor(JobExecutor.class));
     }
 
     private JobStatusCode mapStatusCode(BatchStatusCode batchStatusCode) {
